Moved secret creation service to domain Separated unit tests from API project
